+++ lib/tapioca/compilers/gem/listeners/dynamic_mixins.rb
@@ -18,7 +18,7 @@ module Tapioca
             return if constant.is_a?(Class)
 
             node = event.node
-            mixin_compiler = DynamicMixinCompiler.new(constant)
+            mixin_compiler = Runtime::DynamicMixinCompiler.new(constant)
             mixin_compiler.compile_class_attributes(node)
             dynamic_extends, dynamic_includes = mixin_compiler.compile_mixes_in_class_methods(node)
